2,614 Chinese characters used in given names and nationwide frequency.

2,614 Chinese characters used in given names and nationwide frequency.

data(givenname)

Format

A data frame with 25 variables:

character

character used in given names (in Chinese)

pinyin

pinyin (pronunciation)

bihua

number of strokes in a character

n.male

total counts in male

n.female

total counts in female

name.gender

difference in proportions of a character used by male vs. female

n.1930_1959, n.1960_1969, n.1970_1979, n.1980_1989, n.1990_1999, n.2000_2008

total counts in a birth cohort

ppm.1930_1959, ppm.1960_1969, ppm.1970_1979, ppm.1980_1989, ppm.1990_1999, ppm.2000_2008

proportion (parts per million) in a birth cohort

name.ppm

average ppm (parts per million) across all cohorts

name.uniqueness

name-character uniqueness (in naming practices)

corpus.ppm

proportion (parts per million) in contemporary Chinese corpus

corpus.uniqueness

character-corpus uniqueness (in contemporary Chinese corpus)

name.valence

name valence (positivity of character meaning) (based on subjective ratings from 16 raters, ICC = 0.921)

name.warmth

name warmth/morality (based on subjective ratings from 10 raters, ICC = 0.774)

name.competence

name competence/assertiveness (based on subjective ratings from 10 raters, ICC = 0.712)

Details

https://github.com/psychbruce/ChineseNames